+++ R/effectsize.htest.R
@@ -116,10 +116,16 @@ effectsize.htest <- function(model, type = NULL, verbose = TRUE, ...) {
     return(out)
   } else if (grepl("Fisher's Exact", model$method)) {
     # Fisher's Exact ----
-    stop("Cannot extract effect size from an 'htest' of Fisher's exact test.",
-         "\nTry using 'cramers_v()' or 'phi()' directly.",
-         call. = FALSE
-    )
+    if (is.null(model$estimate))
+      stop("Cannot extract effect size for Fisher's exact test on a table larger than 2x2.",
+           call. = FALSE)
+
+    out <- oddsratio(c(1,2), c(1,2), ci = NULL)
+    out[[1]] <- unname(model$estimate)
+    out$CI <- attr(model$conf.int, "conf.level")
+    out$CI_low <- model$conf.int[1]
+    out$CI_high <- model$conf.int[2]
+    return(out)
   } else if (grepl("Wilcoxon", model$method)) {
     # Wilcoxon ----
     stop("This 'htest' method is not (yet?) supported.", call. = FALSE)